Backend/requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
cloud-sql-python-connector==1.4.3
Flask-Cors==4.0.0
Folium==0.15.0
bcrypt==4.0.1